QR codes have been generated.
QRCreationController#createQRs creates a set of Page objects that
contain QRCodes. It generates these QR Codes in parallel, delegating the
functionality to QRCreator#call. If any Error occurs QRErrorBus will be
filled and queue of handlers will be called. When a QR code is created
it will stored into the Database using the PageDAO.