Refactored and tested QRErrorBus to allow for more polymorphism
In the future, we are going to want to send messages down a message
queue to allow for GUIs to properly handle errors in a multithreaded
environment. Because of this I refactored the main logic that was in
QRErrorBus to be in the abstract AsyncMessageBus class. AsyncMessageBus
will allow us to implement more QRErrorBus like message queues in the
future for our other multithreaded algorithms.