Created the MapReduce pipeline for scanning QR codes
The PDF processing pipeline associates students with assignment pages
by scanning QR codes from the pages. With the integration of Josh's task
and Cam's task and the addition of a parallel java stream I have
implemented this pipeline. See the JavaDoc on
era.uploader.processing.PDFProcessor.process for more detail