Now using JOOQ as ORM and setup a Integration testing framework.
JOOQ is an Object to relational mapping. It allows us to interact with
the database using our domain classes. Our data access objects will be
relying on JOOQ for CRUD operations in the future. It is imperative that
we test all the new DAO code, so I also added some integration testing
related stuff to allow for that.