Updated all_assignments to work with multiple assignments
A grader could accidentally create multiple assignments
resubmitting a pdf twice. Ideally we don't want this to
happen but we have no way of preventing it. I rearranged
my view so that we group by student and the time we
created the assignment, and we only grab the record with
the max created date from that group.