Using JOOQ converts in our database DAO.
The I am using classes in era.uploader.data.convertors in the database
DAOs to convert our model objects to records and the reverse. When we
are converting records to models we grab associated objects (up to one
level in relation) from the database with it. I also made all DAOs
singletons.