Wrote a unit test for CourseDAO#squashMap and added null annotations.
A large part of this next sprint is going to be bugfixes so I decided to
help out Cam and add nullability checks on all database methods. This
will make intellij complain about them. I also noticed that the map
squashing we do in CourseDAO isn't only not tested, but requires some
relatively complex logic. I decided to separate it out into its own
method and write a unit test for it.