Re-created the initial schema and set up flyway
I set up flyway so that if we build locally we don't have
to worry about manually applying SQL scripts, in their
correct order. I also refactored the initial schema to
match our currently schema on the uploader, and to have less,
useless information. Going to migrate the DAOs from the uploader.