Fixed bug where PDF scanning would frozen on a large document.
Most of PDF scanning was being done on another thread. Specifically
the logic that maps qrs to students. However, the splitting up of a
multi-page document wasn't. I had to rework the pdf processing algorithm
to allow for that. I also removed a lot of unused classes and added
documentation.