Fixed some logic bugs that were introduced by Zach's bugfix branch
While Zach's bugfix branch still worked I noticed some dirtiness in our
data that was caused by it. First when we were resolving the course and
we found that it already exists, we just insert all the students on that
course into the database. The problem with that was that we were not
giving _our_ POJO the id of the resolved course, so course_student
records where being inserted with a course_id of 0. Second, we were
inserting duplicate student_ids in course_student basically making
students be counted twice. I fixed that be resolving existing
course_student records. Finally, we weren't truly preserving uniqueness
of courses, because we were always checking semester_id of 0.