Generalized the approach of saving new QR code images.
Dr. Jones wanted us to save QR codes in a different format then
shipping labels, but he likes the option of having shipping labels.
With that in mind I created an abstract class implemented by all
QRSavers that generalizes the saving of QR codes. If Dr. Jones wants
to add a new template label, all the work that is involved is the
math of calculating where to print stuff.