Updated existing documentation and create a service
Because we have to rewrite a lot of QRCreationService and its
name doesn't really correspond to its controller, I decided to
deprecate it and create a new class. This class implements an
AssignmentCreationService interface, so we can change the UI
while we change the backend. I also moved a lot of logic that
are coupled with the with the service layer to be subpackages
of the service layer package.