Added UUIDs to each table on the Uploader side.
Previously we couldn't have multiple uploaders with one server. I have
fixed this by adding UUIDs to each Model that we store in the database.
UUIDs are universally unique so if we use them as primary keys on the
server side we are pretty much guaranteed to not have conflicts amongst
multiple uploaders uploading.