Fix scenario for when we have an offline course and online assignment
Before: If we had an offline course and an assignment that was submitted
for that offline course, we just wouldn't insert the test and failed
silently.

Now: If we have that data scenario the server will send a 409 HTTP
status and the uploader will log an error so the user can see it.